backend/requirements.txt
fastapi
uvicorn
rembg
pillow
python-multipart
onnxruntime-silicon
numpy<2
pillow-heif
vtracer
